+++ CHANGELOG.md
@@ -15,6 +15,9 @@ and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0
 ### Deprecated
 ### Removed
 ### Fixed
+
+- Heisenbugs in `nondet x = oneOf(S)` should not happen (#712)
+
 ### Security
 
 ## v0.7.0 -- 2023-03-09